// src/app/formly-builder/components/app-graph/models/app-graph.models.ts
/* import { ClusterNode, Edge, Layout, MiniMapPosition, Node, NodeDimension } from '@swimlane/ngx-graph';

enum AppGraphCurves {
  Bundle = 'Bundle',
  Cardinal = 'Cardinal',
  Linear = 'Linear',
  Natural = 'Natural',
  Step = 'Step',
  StepAfter = 'Step After',
  StepBefore = 'Step Before'
};

interface AppGraphConfiguration {
  viewSize?: [number, number];
  layout?: string | Layout;
  autoCenter?: boolean;
  autoZoom?: boolean;
  draggingEnabled?: boolean;
  enableZoom?: boolean;
  panningEnabled?: boolean;
  panOnZoom?: boolean;
  showMiniMap?: boolean;
  minZoomLevel?: number;
  maxZoomLevel?: number;
  zoomLevel?: number;
  zoomSpeed?: number;
  miniMapMaxWidth?: number;
  miniMapMaxHeight?: number;
  miniMapPosition?: MiniMapPosition;
}

export {
  AppGraphCurves,
  AppGraphConfiguration,
  ClusterNode,
  Edge,
  Node,
  MiniMapPosition,
  NodeDimension
}
 */